requirements.txt
# Python
unipath==1.1
wheel==0.24.0

# Django
Django==1.9.1
django-debug-toolbar==1.4
django-extensions==1.6.1
djangorestframework==3.3.2

# DB and Models
dj-database-url==0.3.0
django-phonenumber-field==1.0.0
django-solo==1.1.1
psycopg2==2.6
django-admin-sortable2==0.6.2

# Admin
django-jet==0.1.3

# Image handling
Pillow==3.1.0
django-imagekit

# Static Assets
boto
django-assets
django-storages
libsass
whitenoise==2.0.6

# Dev Ops & Deployment
waitress==0.8.10
Werkzeug==0.11.3
